Extra MSVC warning fixes in OGR core
git-svn-id: https://svn.osgeo.org/gdal/trunk@31612 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrcurvecollection.cpp
@@ -327,7 +327,7 @@ OGRErr OGRCurveCollection::exportToWkt( const OGRGeometry* poGeom,
             (*ppszDstText)[nCumulativeLength++] = ',';
 
         /* We must strip the explicit "LINESTRING " prefix */
-        int nSkip = 0;
+        size_t nSkip = 0;
         if( !papoCurves[iGeom]->IsEmpty() &&
             STARTS_WITH_CI(papszGeoms[iGeom], "LINESTRING ") )
         {
